MapTo: move to new package, fix polymorphic type handling
